# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MediaView": {
        "additionalProperties": false,
        "properties": {
          "alt": {
            "title": "Alt",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "alt",
          "created_at"
        ],
        "title": "MediaView",
        "type": "object"
      },
...
      "Product": {
        "additionalProperties": false,
        "properties": {
          "allergens": {
            "minLength": 1,
            "title": "Allergens",
            "type": "string"
          },
          "consumer_care": {
            "minLength": 1,
            "title": "Consumer Care",
            "type": "string"
          },
          "description": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "dietary": {
            "enum": [
              "veg",
              "non-veg"
            ],
            "title": "Dietary",
            "type": "string"
          },
          "ingredients": {
            "minLength": 1,
            "title": "Ingredients",
            "type": "string"
          },
          "manufacturer": {
            "minLength": 1,
            "title": "Manufacturer",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "net_quantity": {
            "minLength": 1,
            "title": "Net Quantity",
            "type": "string"
          },
          "nutrition": {
            "minLength": 1,
            "title": "Nutrition",
            "type": "string"
          },
          "price_paise": {
            "exclusiveMinimum": 0.0,
            "maximum": 100000000.0,
            "title": "Price Paise",
            "type": "integer"
          },
          "shelf_life": {
            "minLength": 1,
            "title": "Shelf Life",
            "type": "string"
          },
          "slug": {
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "description",
          "price_paise",
          "dietary",
          "ingredients",
          "allergens",
          "nutrition",
          "net_quantity",
          "shelf_life",
          "manufacturer",
          "consumer_care"
        ],
        "title": "Product",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Patnam Pakodi Content API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/media": {
      "get": {
        "operationId": "list_media_v1_admin_media_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaView"
                  },
                  "title": "Response List Media V1 Admin Media Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Media"
      },
      "post": {
        "operationId": "upload_media_v1_admin_media_post",
        "parameters": [
          {
            "in": "query",
            "name": "alt",
            "required": true,
            "schema": {
              "maxLength": 300,
              "minLength": 1,
              "title": "Alt",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Media"
      }
    },
...
    "/v1/products/{slug}": {
      "get": {
        "operationId": "getProduct",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Product"
      }
    },
...
  }
}